+++ changelog.md
@@ -20,7 +20,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Fixed
 
--
+- updated packages
 
 ## [7.2.1] - 2026-02-22