+++ changelog.md
@@ -16,17 +16,18 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Changed
 
--
+- secured input form with honeypot
 
 ### Fixed
 
--
+- updated packages
 
 ## [7.2.1] - 2026-02-22
 
 ### Changed
 
 - security changes to input form
+
 ## [7.2.0] - 2026-01-18
 
 ### Added