+++ changelog.md
@@ -16,12 +16,21 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Changed
 
+-
+
+### Fixed
+
+-
+
+## [7.2.2] - 2026-02-22
+
+### Changed
+
 - secured input form with honeypot
 
 ### Fixed
 
 - updated packages
-
 ## [7.2.1] - 2026-02-22
 
 ### Changed